// 05-12-23 UATsrc/src/app/models/common.Model.ts
export interface IPolicyAndPremiumStats {
  NoOfPolicies: number;
  TotalPremiumPerYr: number;
}

export class IPolicyRecord {
  name: string;
  mobileNo: number;
  regNo: string;
  insurer: string;
  premium: number;
  paymentDate: string;
  policyNo: string;
  vehicleNo?: string;
  DownloadLinkURL?: string;
  CommisionAmt?: Number;
  InpectionCheckStarus?: string;

  constructor(
    name: string,
    mobile: number,
    regNo: string,
    insurer: string,
    premium: number,
    paymentdate: string,
    policyNo: string,
    vehicleNo?: string | undefined,
    DownloadLinkURL?: string | undefined,
    CommisionAmt?: Number | undefined,
    InpectionCheckStarus?: string | undefined,


  ) {
    this.name = name;
    this.mobileNo = mobile;
    this.regNo = regNo;
    this.insurer = insurer;
    this.premium = premium;
    this.paymentDate = paymentdate;
    this.policyNo = policyNo;
    this.vehicleNo = vehicleNo;
    this.DownloadLinkURL = DownloadLinkURL;
    this.CommisionAmt = CommisionAmt;
    this.InpectionCheckStarus = InpectionCheckStarus;
  }
}
// ...
export interface IPolicyPDFDocDownload {
  PolicyNo: string;
  ProposalPolicyNumber: string;
  ReturnKeyOfPayment: string;
  CompanyCode: string;
  ProposalProduct: string;
  ApplicationNo: string;
  ApplicationNoOdp: string;
  RegistrationNo: string;
  RegistrationNoOdp: string;
  ProposalOrderNo: string;
  ComprehensiveThirdPartyCode: string;
  SubCateCode: number
  InsuranceCateCode: number
  TransactionNo: string
}